sparse voxel octree lods, ram optimizations, chunk visualization, camera improvements and culling

+++ src/voxel/Voxel.hpp
@@ -3,7 +3,7 @@
 #include "raylib.h"
 #include <cstdint>
 
-enum BlockType : uint16_t {
+enum BlockType : uint8_t {
     BLOCK_AIR = 0,
     BLOCK_GRASS = 1,
     BLOCK_DIRT = 2,
